Fix retyping card_last4 as String
The type was wrong for card_last4 as it was originally typed as a short.
The short type did not account that the card_last4 would sometimes start
with a zero causing the short to drop the zero. The variable has been
retyped to String in order to preserve leading zeros.

+++ backend/payment_reconciliation/src/main/java/com/platinumrelations/interview/payment_reconciliation/reconciliation/service/engine/WideWindowCategorizer.java
@@ -16,7 +16,7 @@ import java.util.List;
 @Service
 class WideWindowCategorizer {
 
-    WideWindowCategorizer(@Value("${app.custom.wide-window.max-days}") Short windowMaxDays, @Value("${app.custom.holidays}") List<String> holidays){
+    WideWindowCategorizer(@Value("${app.custom.wide-window.max-days}") int windowMaxDays, @Value("${app.custom.holidays}") List<String> holidays){
         this.windowMaxDays = windowMaxDays;
 
         this.holidays = new ArrayList<>();
@@ -32,7 +32,7 @@ class WideWindowCategorizer {
         }
     }
 
-    private final short windowMaxDays;
+    private final int windowMaxDays;
     private final List<LocalDate> holidays;
 
     boolean hasWideWindow(InternalTransaction it, ProcessorSettlement ps){
